Bug fixed: any user could create plays for any other user.
This was because req.isAuthenticated stored in req.user the identity of
the authenticated user.
Then, play.userMiddleware() stored in req.user the user
specified in the URL for the play, overwriting the authenticated user identity.
So, the ID of the owner of the play and the ID of the authenticated user
would always be identical.

Now req.isAuthenticated stores the UID in req.user, and
play.userMiddleware uses req.owner.